requirements.txt
llama-index>=0.10.1
llama-index-core>=0.10.1,<0.11.0
llama-index-llms-openai>=0.1.7
llama-index-embeddings-openai>=0.1.7
langgraph==0.2.60
python-dotenv==1.0.0
google-search-results==2.4.2
pandas==2.1.4
plotly==5.18.0
dash==2.14.2
pydantic>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10